Secured views with superuser rights

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import get_object_or_404, render, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -65,8 +66,13 @@ def all_products(request):
 	return render(request, 'products/products.html', context)
 
 
+@login_required
 def add_product(request):
 	""" For admin use to add a product to the store """
+	if not request.user.is_superuser:
+		messages.error(request, 'Sorry, your are not allowed to add a product, that is for store only.')
+		return redirect(reverse('home'))
+
 	if request.method == "POST":
 		form = ProductForm(request.POST)
 		if form.is_valid():
@@ -86,8 +92,13 @@ def add_product(request):
 	return render(request, template, context)
 
 
+@login_required
 def update_product(request, product_id):
 	""" For admin use to update a product in the store """
+	if not request.user.is_superuser:
+		messages.error(request, 'Sorry, your are not allowed to update a product, that is for store only.')
+		return redirect(reverse('home'))
+	
 	product = get_object_or_404(Product, pk=product_id)
 	if request.method == 'POST':
 		form = ProductForm(request.POST, instance=product)
@@ -109,8 +120,13 @@ def update_product(request, product_id):
 	return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
 	""" For admin use to delete a product in the store """
+	if not request.user.is_superuser:
+		messages.error(request, 'Sorry, your are not allowed to delete a product, that is for store only.')
+		return redirect(reverse('home'))
+	
 	product = get_object_or_404(Product, pk=product_id)
 	product.delete()
 	messages.success(request, 'Product has been deleted succesfully!')
